requirements.txt
Celery
Django
django-celery-beat 
django-celery-results
python-decouple
redis
selenium